+++ Source/EventFlow.Autofac.Tests/UnitTests/Aggregates/AutofacAggregateFactoryTests.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -27,6 +27,7 @@ using EventFlow.Aggregates;
 using EventFlow.Autofac.Extensions;
 using EventFlow.Configuration;
 using EventFlow.Extensions;
+using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Aggregates;
 using FluentAssertions;
 using NUnit.Framework;
@@ -34,6 +35,7 @@ using NUnit.Framework;
 namespace EventFlow.Autofac.Tests.UnitTests.Aggregates
 {
     [TestFixture]
+    [Category(Categories.Scenario)]
     public class AutofacAggregateFactoryTests
     {
         [Test]