+++ Source/EventFlow.Hangfire.Tests/Integration/HangfireJobFlow.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -40,6 +40,7 @@ using EventFlow.TestHelpers.Aggregates.ValueObjects;
 
 namespace EventFlow.Hangfire.Tests.Integration
 {
+    [Category(Categories.Integration)]
     public class HangfireJobFlow : Test
     {
         private IMsSqlDatabase _msSqlDatabase;