+++ Source/EventFlow.ReadStores.MsSql/MssqlReadModelStore.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -34,6 +34,7 @@ using EventFlow.Extensions;
 using EventFlow.Logs;
 using EventFlow.MsSql;
 using EventFlow.ReadStores.MsSql.Attributes;
+using EventFlow.Sql.ReadModels;
 
 #pragma warning disable 618