+++ Source/EventFlow.Tests/IntegrationTests/EventStores/FilesEventStoreTests.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -26,11 +26,13 @@ using System.IO;
 using EventFlow.Configuration;
 using EventFlow.EventStores.Files;
 using EventFlow.Extensions;
+using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Suites;
 using NUnit.Framework;
 
 namespace EventFlow.Tests.IntegrationTests.EventStores
 {
+    [Category(Categories.Integration)]
     public class FilesEventStoreTests : TestSuiteForEventStore
     {
         private IFilesEventStoreConfiguration _configuration;