+++ Source/EventFlow.Tests/IntegrationTests/ReadStores/InMemoryReadModelStoreTests.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -26,13 +26,16 @@ using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Configuration;
 using EventFlow.Extensions;
+using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Aggregates.Entities;
 using EventFlow.TestHelpers.Suites;
 using EventFlow.Tests.IntegrationTests.ReadStores.QueryHandlers;
 using EventFlow.Tests.IntegrationTests.ReadStores.ReadModels;
+using NUnit.Framework;
 
 namespace EventFlow.Tests.IntegrationTests.ReadStores
 {
+    [Category(Categories.Scenario)]
     public class InMemoryReadModelStoreTests : TestSuiteForReadModelStore
     {
         protected override IRootResolver CreateRootResolver(IEventFlowOptions eventFlowOptions)