+++ Source/EventFlow.Tests/UnitTests/Configuration/Registrations/AutofacServiceRegistrationTests.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -28,6 +28,7 @@ using NUnit.Framework;
 
 namespace EventFlow.Tests.UnitTests.Configuration.Registrations
 {
+    [Category(Categories.Unit)]
     public class AutofacServiceRegistrationTests : Test
     {
         // ReSharper disable ClassNeverInstantiated.Local