+++ Source/EventFlow.Tests/UnitTests/Core/RetryDelayTests.cs
@@ -1,0 +1,53 @@
+﻿// The MIT License (MIT)
+// 
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
+// https://github.com/rasmus/EventFlow
+// 
+// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to
+// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+// the Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+// FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+// COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+// IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+// CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+using System;
+using EventFlow.Core;
+using FluentAssertions;
+using NUnit.Framework;
+
+namespace EventFlow.Tests.UnitTests.Core
+{
+    public class RetryDelayTests
+    {
+        [Test]
+        [Repeat(42)] // 42 is indeed that number of times to run a test to ensure its always true
+        public void PickDelay_IsWithinBounds()
+        {
+            // Arrange
+            const double max = 1000;
+            const double min = 500;
+            var sut = RetryDelay.Between(
+                TimeSpan.FromMilliseconds(min),
+                TimeSpan.FromMilliseconds(max));
+
+            // Act
+            var delay = sut.PickDelay();
+
+            // Assert
+            delay.TotalMilliseconds.Should().BeGreaterOrEqualTo(min);
+            delay.TotalMilliseconds.Should().BeLessOrEqualTo(max);
+        }
+    }
+}