+++ Source/EventFlow/Extensions/DisposableExtensions.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015 Rasmus Mikkelsen
-// Copyright (c) 2015 eBay Software Foundation
+// Copyright (c) 2015-2016 Rasmus Mikkelsen
+// Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -28,6 +28,13 @@ namespace EventFlow.Extensions
 {
     public static class DisposableExtensions
     {
+        public static void DisposeSafe(
+            this IDisposable disposable,
+            string message)
+        {
+            DisposeSafe(disposable, new ConsoleLog(), message);
+        }
+
         public static void DisposeSafe(
             this IDisposable disposable, 
             ILog log,