+++ RELEASE_NOTES.md
@@ -1,4 +1,20 @@
-### New in 0.25 (not released yet)
+### New in 0.26 (not released yet)
+
+ * Breaking: Renamed `MssqlMigrationException` to `SqlMigrationException`
+ * Breaking: Renamed `SqlErrorRetryStrategy` to `MsSqlErrorRetryStrategy`
+   as its MSSQL specific
+ * Breaking: The NuGet package `Dapper` is no longer IL merged with the package
+   `EventFlow.MsSql` but is now listed as a NuGet dependency. The current
+   version used by EventFlow is `v1.42`
+ * New: Introduced the NuGet package `EventFlow.SQLite` that adds event store
+   support for SQLite databases, both as event store and read model store
+ * New: Introduced the NuGet package `EventFlow.Sql` as shared package for
+   EventFlow packages that uses SQL
+ * New: Its now possible to configure the retry delay for MSSQL transient
+   errors using the new `IMsSqlConfiguration.SetTransientRetryDelay`. The
+   default is a random delay between 50 and 100 milliseconds
+
+### New in 0.25.1695 (released 2016-02-15)
 
 * Fixed: Deadlock in `AsyncHelper` if e.g. an exception caused no `async` tasks
   to be scheduled. The `AsyncHelper` is used by EventFlow to expose non-`async`